Added test to validate TextChoice in an issue design. Moved other domain tests into the import/export tests. Added comments.
