+++ src/org/labkey/test/tests/TextChoiceTest.java
@@ -8,7 +8,6 @@ import org.labkey.test.pages.ReactAssayDesignerPage;
 import org.labkey.test.params.FieldDefinition;
 import org.labkey.test.params.experiment.SampleTypeDefinition;
 import org.labkey.test.util.DataRegionTable;
-import org.labkey.test.util.ListHelper;
 import org.labkey.test.util.TestDataGenerator;
 import org.labkey.test.util.exp.SampleTypeAPIHelper;
 
@@ -31,7 +30,7 @@ public abstract class TextChoiceTest extends BaseWebDriverTest
     @Override
     public List<String> getAssociatedModules()
     {
-        return Arrays.asList("experiment");
+        return Arrays.asList("experiment", "issues");
     }
 
     // Sample Type field names, values etc...
@@ -50,26 +49,16 @@ public abstract class TextChoiceTest extends BaseWebDriverTest
 
     protected static final String BATCH_TC_FIELD = "Batch_TC_Field";
     protected static final List<String> BATCH_FIELD_VALUES = List.of("B1", "B2", "B3");
-    protected static final String BATCH_VALUE = "B2";
+    protected static final String BATCH_VALUE = BATCH_FIELD_VALUES.get(1);
 
     protected static final String RUN_TC_FIELD = "Run_TC_Field";
     protected static final List<String> RUN_FIELD_VALUES = List.of("RN1", "RN2", "RN3", "RN4", "RN5");
-    protected static final String RUN_VALUE = "RN2";
+    protected static final String RUN_VALUE = RUN_FIELD_VALUES.get(1);
 
     protected static final String RESULT_TC_FIELD = "Result_TC_Field";
     protected static final List<String> RESULT_FIELD_VALUES = List.of("RS1", "RS2", "RS3", "RS4", "RS5", "RS6", "RS7", "RS8", "RS9", "RS10");
 
     protected static final String RESULT_SAMPLE_FIELD = "Sample";
-
-    // List field name, values etc...
-    protected static final String LIST_NAME = "Simple_TC_List";
-    protected static final String LIST_TC_FIELD = "LTC_Field";
-    protected static final String LIST_TEXT_FIELD = "Str";
-    protected static final List<String> LIST_VALUES = Arrays.asList("L1", "L2", "L3", "L4");
-    protected static final List<Map<String, String>> LIST_DATA = Arrays.asList(
-            Map.of(LIST_TC_FIELD, LIST_VALUES.get(0), LIST_TEXT_FIELD, "My"),
-            Map.of(LIST_TC_FIELD, LIST_VALUES.get(1), LIST_TEXT_FIELD, "Name"),
-            Map.of(LIST_TC_FIELD, LIST_VALUES.get(2), LIST_TEXT_FIELD, "Is"));
 
     /**
      * Create a sample type using the 'default values'. Does no validation of the design.
@@ -77,7 +66,7 @@ public abstract class TextChoiceTest extends BaseWebDriverTest
      * @throws IOException Can be thrown by the create command.
      * @throws CommandException Can be thrown by the create command.
      */
-    protected void createSampleType() throws IOException, CommandException
+    protected void createDefaultSampleTypeWithTextChoice() throws IOException, CommandException
     {
         log(String.format("Create a new sample type named '%s'.", ST_NAME));
         SampleTypeDefinition sampleTypeDefinition = new SampleTypeDefinition(ST_NAME);
@@ -127,7 +116,7 @@ public abstract class TextChoiceTest extends BaseWebDriverTest
     /**
      * Helper to create an assay design using the 'default values'. It does no validation of the design.
      */
-    protected void createAssayDesign()
+    protected void createDefaultAssayDesignWithTextChoice()
     {
         goToManageAssays();
         ReactAssayDesignerPage assayDesignerPage = _assayHelper.createAssayDesign("General", ASSAY_NAME)
@@ -205,28 +194,6 @@ public abstract class TextChoiceTest extends BaseWebDriverTest
             checker().verifyEquals(String.format("Result row not as expected for sample '%s'.", entry.getKey()), expectedData, rowData);
         }
 
-    }
-
-    /**
-     * Simple helper to create a list using the 'default values'. It does no validation of the list design.
-     */
-    protected void createList()
-    {
-
-        FieldDefinition tcField = new FieldDefinition(LIST_TC_FIELD, FieldDefinition.ColumnType.TextChoice);
-        tcField.setTextChoiceValues(LIST_VALUES);
-
-        FieldDefinition txtField = new FieldDefinition(LIST_TEXT_FIELD, FieldDefinition.ColumnType.String);
-
-        _listHelper.createList(getCurrentContainerPath(), LIST_NAME, ListHelper.ListColumnType.AutoInteger, "Key", tcField, txtField);
-
-        StringBuilder sb = new StringBuilder();
-        sb.append(String.format("%s\t%s\n", LIST_TC_FIELD, LIST_TEXT_FIELD));
-        for(Map<String, String> row : LIST_DATA)
-        {
-            sb.append(String.format("%s\t%s\n", row.get(LIST_TC_FIELD), row.get(LIST_TEXT_FIELD)));
-        }
-        _listHelper.uploadData(sb.toString());
     }
 
 }
